services/web/tasks/requirements.txt
Flask-RESTful==0.3.9
celery==5.2.3
Flask==2.0.3
redis==4.1.4
pydub==0.25.1
requests==2.28.1
google-cloud-storage